Safari-609.1.11.1/LayoutTests/http/tests/misc/char-encoding-scsu-blacklisted-expected.txt

Test that character set SCSU is not supported and we default to ISO-8859-1. With support for SCSU the iframe renders a Cyrillic string. Without it renders garbage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS charset is windows-1252
PASS successfullyParsed is true

TEST COMPLETE
